requirements.txt
numpy
pillow
opencv-python

torch
torchvision
ultralytics

streamlit
pandas
altair

reportlab
